% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{align}
\alias{align}
\title{Alignment with kernel gaussian density}
\usage{
align(peakTab, ppmGroup = 70, dmzGroup = 0.001)
}
\arguments{
\item{peakTab}{table with comlumn : mass, quantification, and groups 
number to aligned}

\item{ppmGroup}{width of sub group created beafore density estimation in ppm}

\item{dmzGroup}{width of sub group created beafore density estimation in Da}
}
\value{
A list containing groups formed by alignment.
}
\description{
Alignment with kernel gaussian density
}
\keyword{internal}
