% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignement.R
\name{alignExpirations}
\alias{alignExpirations}
\title{Align mass tables of different expirations}
\usage{
alignExpirations(PeakMat, ppmGroup = 70, fracGroup = 0.8, dmzGroup = 0.001)
}
\arguments{
\item{PeakMat}{A peakMatrix, with at least column named 
"mz" correspond to masses, and an other named "group" correspond to the expiration number.}

\item{ppmGroup}{the ppm larger of a group}

\item{fracGroup}{We will keep variables only present in \code{fracGroup} of expirations}

\item{dmzGroup}{difference of mz of a group for little mz}
}
\value{
A list of data.table with same length as PeakList input, but the alignment between 
expirations is done: each matrix contains columns "mz", "delta_mz" and "ppb", "fracExp" and "background"
}
\description{
Align mass tables of different expirations
}
