% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_Methods.R
\name{calibrationFun}
\alias{calibrationFun}
\title{calibration function}
\usage{
calibrationFun(sp, mz, mzCalibRef, calibCoef, tol)
}
\arguments{
\item{sp}{spectrum}

\item{mz}{mass axis}

\item{mzCalibRef}{masses of know reference peaks}

\item{calibCoef}{coeficient of the previous calibration}

\item{tol}{maximum error tolarated in ppm}
}
\value{
list
}
\description{
Performs calibration on sp with mzCalibRef reference masses and mzToTofFunc as previous 
calibration function
}
