% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{calibShiftCorr}
\alias{calibShiftCorr}
\title{corretcion of the tof shift}
\usage{
calibShiftCorr(rawM, raw, peak, listCalib)
}
\arguments{
\item{rawM}{an EIC}

\item{raw}{ptrRaw object (replaec by the coef calib)}

\item{peak}{mz peak in rawM}

\item{listCalib}{mutlicalib list}
}
\value{
the rawM corrected
}
\description{
corretcion of the tof shift
}
