% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{calibShitEstimate}
\alias{calibShitEstimate}
\title{Estimate the calibration for precise mz}
\usage{
calibShitEstimate(raw, peak, listCalib)
}
\arguments{
\item{raw}{a ptrRaw object}

\item{peak}{a data.frame with a column named "Mz". The Mz of the VOC detected}

\item{listCalib}{a list containing calibration coeficient for different time periods:
the output of multicalib function}
}
\value{
a vector with mz shift values in Da for eaf mz of peak list
}
\description{
Estimate the calibration for precise mz
}
