% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{changeTimeLimits}
\alias{changeTimeLimits}
\title{Shinny app to modify expiration.
 
This function run a shiny app, where you can check the automatic expirations/head spaces detection, and modify it.}
\usage{
changeTimeLimits(ptrSet)
}
\arguments{
\item{ptrSet}{a ptrSet}
}
\value{
the ptrSet object modified
}
\description{
Shinny app to modify expiration.
 
This function run a shiny app, where you can check the automatic expirations/head spaces detection, and modify it.
}
\examples{
library(ptairData)
directory <- system.file("extdata/exhaledAir",  package = "ptairData")
ptrSet <- createPtrSet(directory,setName="ptrSet",mzCalibRef=c(21.022,59.049),fracMaxTIC=0.8)
\dontrun{ptrSet <- changeTimeLimits(ptrSet)}
}
