% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{checkSet}
\alias{checkSet}
\title{check all files in a directory}
\usage{
checkSet(files, mzCalibRef, fracMaxTIC, mzBreathTracer)
}
\arguments{
\item{files}{list of full names files to chexk}

\item{mzCalibRef}{Vector of accurate mass values of intensive peaks and 'unique' in a 
nominal mass interval (without overlapping)}

\item{fracMaxTIC}{timeLimits argument}
}
\value{
list containing in ptrSet object slot
}
\description{
main function of createPtrSet: opens each files, performs calibration, calculates TIC boundaries and
quantify primary ion .
}
