% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{computeTemporalFile}
\alias{computeTemporalFile}
\title{extract quantity over time of VOC in a pre-defined peak list}
\usage{
computeTemporalFile(
  raw,
  peak,
  indTimeLim,
  timeCalib = 20,
  deconvMethod = deconv2d2linearIndependant,
  ...
)
}
\arguments{
\item{raw}{ptrRaw object}

\item{peak}{a data.frame with a column named "Mz". The Mz of the VOC detected}

\item{indTimeLim}{expiration and and bg limit}

\item{timeCalib}{duration in second of every calibration}

\item{deconvMethod}{the deconvolution 2d method. Should be "deconv2d2linearIndependant",
"deconv2d2LinearCoupled" or "deconv2d2NonlinearIndependant"}
}
\value{
list containing the matrix with all temporal profile (matPeak) 
of VOC and a list of all raw EIC (EIClist)
}
\description{
extract quantity over time of VOC in a pre-defined peak list
}
