% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{cumulative_fit_function}
\alias{cumulative_fit_function}
\title{Create cumulative function fit}
\usage{
cumulative_fit_function(fit_function_str, par_var_str, par_fix_str, n.peak)
}
\arguments{
\item{fit_function_str}{fit function who will be use in character}

\item{par_var_str}{parameters of fit function who change with the peak in a 
vector of character}

\item{par_fix_str}{parameters of fit function independent of the peak in a 
vector of character}

\item{n.peak}{number of peak}
}
\value{
a list: \cr
  \code{init.names}: names of paramters for the initialization \cr
  \code{func.eval}: function who will be fitted
}
\description{
Create cumulative function fit
}
