% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{determinePeakShape}
\alias{determinePeakShape}
\title{Determine peak shape from raw data}
\usage{
determinePeakShape(
  sp,
  mz,
  massRef = c(21.022, 29.013424, 41.03858, 75.04406, 203.943, 330.8495),
  plotShape = FALSE
)
}
\arguments{
\item{sp}{vector with spectrum values where the peak shape will be extimated}

\item{mz}{vectr with mass axis}

\item{massRef}{the accurate mass of the reference peak (for example the same that used for calibration)}

\item{plotShape}{if true plot each reference peak and the average peak (the peak shape)}
}
\value{
A list of two vectors which are the reference peak normalized tof and intensity.
}
\description{
This function use the method descibe by average and al 2013, for determine a peak 
shape from the raw data : \cr 
$peak_i(Delta_i,A_i,t_i) = interpolation(x= tof.ref * Delta_i + t_i,y = A_i * 
peak.ref, xout= TOF_i )$ where peak.ref and tof.ref are peaks reference use for mass calibration.
}
