% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.ptrms_vocRef}
\alias{.ptrms_vocRef}
\title{.vocRef}
\usage{
.ptrms_vocRef(mzRgeVn)
}
\arguments{
\item{mzRgeVn}{range of mz values}
}
\value{
data frame with the reference VOCs
}
\description{
reference VOC
}
