% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_Methods.R
\name{estimateResol}
\alias{estimateResol}
\title{estimate the resolution of a ptrRaw with reference calibration masses}
\usage{
estimateResol(calibMassRef, calibSpectr)
}
\arguments{
\item{calibMassRef}{references mass for etsimate resolution}

\item{calibSpectr}{list of mz and spectrum of reference masses}
}
\description{
estimate the resolution of a ptrRaw with reference calibration masses
}
