% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exhaledPtrset}
\alias{exhaledPtrset}
\title{PtrSet of the exhaled air rdhf 5 files}
\format{
ptrSet object
}
\usage{
exhaledPtrset
}
\description{
a ptrSet object, created by the \code{ptairMS::createPtrSet} function, containing processed
exhaled air dataset (files present in the inst/extdata/exhaledAir) by applying \code{ptairMS::detectPeak} function
}
\keyword{datasets}
