% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{exportSampleMetada}
\alias{exportSampleMetada}
\title{export sampleMetadata}
\usage{
exportSampleMetada(set, saveFile)
}
\arguments{
\item{set}{a ptrSet object}

\item{saveFile}{a file path in tsv extension where the data.frame will be exported}
}
\value{
nothing
}
\description{
export sampleMetadata
}
\examples{
library(ptairData)
directory <- system.file("extdata/mycobacteria",  package = "ptairData")
mycobacteria <- createPtrSet(dir= directory, setName="mycobacteria",
mzCalibRef= c(21.022, 59.049141))
saveFile<-file.path(directory,"sampleMetadata.tsv")
#exportSampleMetada(mycobacteria,saveFile)
}
