% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{findIsotope}
\alias{findIsotope}
\title{Isotope detection and validation of isotope clusters in PTR-TOF-MS peak table}
\usage{
findIsotope(eSet, ppm = 100)
}
\arguments{
\item{eSet}{an expression set of PTR-TOF-MS data aligned}

\item{ppm}{presision for the mass matching}
}
\value{
an expresion with the column isotope added in teh features data
}
\description{
This function identify possible isotope cluster of C13,O17 and O18 atomes after alignment. It writes in the 
features data of the expressionSet (Biobase::fData()) the isotope m/z in the "isotope" column at the ligne 
of the more intenisf peak of the clusters.
}
\examples{
data(exhaledPtrset)
exhaled.eset <- alignSamples(exhaledPtrset ,fracGroup=0.9,pValGreaterThres=0.05)
exhaled.eset<- impute(exhaled.eset,exhaledPtrset )
exhaled.eset <-findIsotope(exhaled.eset)
Biobase::fData(exhaled.eset)[,"isotope",drop=FALSE]
}
