% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{fit_averagePeak}
\alias{fit_averagePeak}
\title{Fit peak with average function}
\usage{
fit_averagePeak(initTof, l.shape, sp, bin, lower.cons, upper.cons)
}
\arguments{
\item{initTof}{list of initialisation in tof}

\item{l.shape}{peak shape average}

\item{sp}{spectrum}

\item{bin}{tof axis}

\item{lower.cons}{lower constrain for fit}

\item{upper.cons}{upper constrain for fit}
}
\value{
list with fit information
}
\description{
Fit peak with average function
}
