% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{fit_averagePeak_function}
\alias{fit_averagePeak_function}
\title{fit function average}
\usage{
fit_averagePeak_function(t, delta, h, intervRef, peakShape, bin)
}
\arguments{
\item{t}{tof center of peak}

\item{delta}{FWHM of peak}

\item{h}{peak height}

\item{intervRef}{reference intreval for peak shape}

\item{peakShape}{peak shape estimated on \code{intervalRef}}

\item{bin}{bin interval of peak will be fitted}
}
\value{
peak function made on an average of reference peaks normalized
}
\description{
fit function average
}
