% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{fit_sech2}
\alias{fit_sech2}
\title{Fit peak with Sech2 function}
\usage{
fit_sech2(initMz, sp, mz.i, lower.cons, upper.cons)
}
\arguments{
\item{initMz}{list of initial parmeter (mz,delta,h)}

\item{sp}{spectrum}

\item{mz.i}{mass spectrum}

\item{lower.cons}{lower constrain of fit}

\item{upper.cons}{upper constrain of fit}
}
\description{
Fit peak with Sech2 function
}
