% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{fitPeak}
\alias{fitPeak}
\title{Fit peak with Sech2 function}
\usage{
fitPeak(initMz, sp, mz.i, lower.cons, upper.cons, funcName)
}
\arguments{
\item{initMz}{list of initial parmeter (mz,delta,h)}

\item{sp}{spectrum}

\item{mz.i}{mass spectrum}

\item{lower.cons}{lower constrain of fit}

\item{upper.cons}{upper constrain of fit}

\item{funcName}{peak function name, also defined, with 4 parameters: peak center , FWHM at left,  FWHM at right, heigh}
}
\value{
list with fit result
}
\description{
Fit peak with Sech2 function
}
