% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{formula2mass}
\alias{formula2mass}
\title{Compute exact mass.}
\usage{
formula2mass(formula.vc, protonize.l = TRUE)
}
\arguments{
\item{formula.vc}{Vector of molecular formulas.}

\item{protonize.l}{Should a proton be added to the formula?}
}
\value{
Vector of the corresponding (protonated) masses.
}
\description{
Compute exact mass from an elemental formula
}
\examples{
formula2mass("CO2")
}
