% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{getDirectory}
\alias{getDirectory}
\title{get the files directory of a ptrSet}
\usage{
getDirectory(ptrSet)
}
\arguments{
\item{ptrSet}{ptrSet object}
}
\value{
the directory in absolute path as character
}
\description{
get the files directory of a ptrSet
}
\examples{
library(ptairData)
dirRaw <- system.file("extdata/exhaledAir", package = "ptairData")
exhaledPtrset <- createPtrSet(dir=dirRaw, setName="exhaledPtrset", 
mzCalibRef = c(21.022, 60.0525), fracMaxTIC = 0.7, saveDir = NULL )
getDirectory(exhaledPtrset )
}
