% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{importSampleMetadata}
\alias{importSampleMetadata}
\title{import a sampleMetadata from a tsv file to a ptrSet object}
\usage{
importSampleMetadata(set, file)
}
\arguments{
\item{set}{a ptrSet object}

\item{file}{a tsv file contains the sample metada to import, with all 
file names in row name 
(the fist column on th excel).}
}
\value{
a ptrSet with th enew sample Metadata
}
\description{
import a sampleMetadata from a tsv file to a ptrSet object
}
\examples{
library(ptairData)
dirRaw <- system.file("extdata/exhaledAir", package = "ptairData")
exhaledPtrset <- createPtrSet(dir=dirRaw, setName="exhaledPtrset", 
mzCalibRef = c(21.022, 60.0525), fracMaxTIC = 0.7, saveDir = NULL )
saveFile<-file.path(getwd(),'sampleMetadata.tsv')
#exportSampleMetada(exhaledPtrset ,saveFile)
#exhaledPtrset<-importSampleMetadata(exhaledPtrset ,saveFile)
}
