% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{initializeFit}
\alias{initializeFit}
\title{initialization for apply fit function in the spectrum}
\usage{
initializeFit(
  i,
  sp.i.fit,
  sp.i,
  mz.i,
  mzToTof,
  minpeakheight,
  noiseacf,
  ppmPeakMinSep,
  daSeparation,
  d,
  plotAll,
  c
)
}
\arguments{
\item{i}{the nominal mass}

\item{sp.i.fit}{the vector who will be fetted (spectrum pf residual)}

\item{sp.i}{the spectrum around a nominal mass}

\item{mz.i}{the mass vector around a nominal mass}

\item{mzToTof}{the function for convert mz to Tof}

\item{minpeakheight}{the minimum peak intensity}

\item{noiseacf}{aytocorelation of the noise}

\item{ppmPeakMinSep}{the minimum distance between two peeks in ppm}

\item{daSeparation}{the minimum distance between two peeks in da}

\item{d}{the degree of savitzky golay filter}

\item{plotAll}{bollean if TRUE, it plot all the initialiation step}

\item{c}{the number of current itteration}
}
\value{
a list with fit input
}
\description{
initialization for apply fit function in the spectrum
}
