% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{multiCalib}
\alias{multiCalib}
\title{Performs a calibration every x time step}
\usage{
multiCalib(raw, step)
}
\arguments{
\item{raw}{a ptrRaw object already calibrate}

\item{step}{number of time step}
}
\value{
a list that each element contains: index time, coef calibration, and the function to convert mz to tof
}
\description{
Performs a calibration every x time step
}
