% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mycobacteriaSet}
\alias{mycobacteriaSet}
\title{PtrSet of the mycrobacteria h5 file}
\format{
a ptrSet object
}
\usage{
mycobacteriaSet
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
