% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{OptimalWindowsSG}
\alias{OptimalWindowsSG}
\title{Find optimal window's size for Savitzky Golay filter}
\usage{
OptimalWindowsSG(sp, noiseacf, d = 3)
}
\arguments{
\item{sp}{the array of specrtum values}

\item{noiseacf}{autocorrelation of the noise}

\item{d}{the degree of Savitzky Golay filter}
}
\value{
the optimal size of Savitzky Golay filter's windows
}
\description{
Find optimal window's size for Savitzky Golay filter
}
\keyword{internal}
