% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{peakListNominalMass}
\alias{peakListNominalMass}
\title{Detect peak in a single nominal mass, same parameter as peakList}
\usage{
peakListNominalMass(
  i,
  mz,
  sp,
  ppmPeakMinSep = 130,
  calibCoef,
  resMinMeanMax = c(300, 5000, 8000),
  minPeakDetect = 10,
  fitFunc = "sech2",
  maxIter = 2,
  R2min = 0.995,
  autocorNoiseMax = 0.3,
  plotFinal = FALSE,
  plotAll = FALSE,
  thNoiseRate = 1.1,
  thIntensityRate = 0.01,
  countFacFWHM = 10,
  daSeparation = 0.001,
  d = 3,
  windowSize = 0.4,
  l.shape = NULL
)
}
\arguments{
\item{i}{the nominal mass}

\item{mz}{peak mass axis}

\item{sp}{peak spectrum}

\item{ppmPeakMinSep}{the minimum distance between two peeks in ppm}

\item{calibCoef}{calibration coeficient a and b}

\item{minPeakDetect}{the minimum intenisty for peaks detection. The final threshold for peak detection
will be : max ( \code{minPeakDetect} , thresholdNoise ). The thresholdNoise correspond to
 max(\code{thNoiseRate} * max( noise around the nominal mass), \code{thIntensityRate} * 
 max( intenisty in the nominal mass). The noise around the nominal mass correspond : 
 \code{[m-windowSize-0.2,m-windowSize]U[m+windowSize,m+WindowSize+0.2]}.}

\item{fitFunc}{the function for the quantification of Peak, should be average or Sech2}

\item{maxIter}{maximum ittertion of residual analysis}

\item{autocorNoiseMax}{the autocorelation threshold for Optimal windows Savitzky Golay 
filter in \code{OptimalWindowSG} ptairMS function. See \code{?OptimalWindowSG}}

\item{plotFinal}{boolean. If TRUE, plot the spectrum for all nominal masses, with the final fitted peaks}

\item{plotAll}{boolean. Tf TRUE, plot all step to get the final fitted peaks}

\item{thNoiseRate}{The rate who is multiplie by the max noise intensity}

\item{thIntensityRate}{The rate who is mutluplie by the max signal intensity}

\item{countFacFWHM}{integer. We will sum the fitted peaks on a windows's size of countFacFWHM * FWHM, 
centered in the mass peak center.}

\item{daSeparation}{the minimum distance between two peeks in Da for nominal mass < 17.}

\item{d}{the degree for the \code{Savitzky Golay} filtrer}

\item{windowSize}{peaks will be detected only around  m - windowSize ; m + windowSize, for all 
m in \code{mzNominal}}

\item{l.shape}{peak shape averae}

\item{resMinMaxMean}{vector with resolution min, resolution Mean, and resolution max of the PTR}
}
\value{
a list that contains the peak list of nomminal mass i, and information to plot the peaks detected
}
\description{
Detect peak in a single nominal mass, same parameter as peakList
}
