% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{plot,ptrSet,ANY-method}
\alias{plot,ptrSet,ANY-method}
\alias{plot.ptrSet}
\alias{plot,ptrSet-method}
\title{ptrSet object}
\usage{
\S4method{plot}{ptrSet,ANY}(x, y, typePlot = "")
}
\arguments{
\item{x}{a ptrSet object}

\item{y}{not use}

\item{typePlot}{could be : \code{calibError}, \code{resolution},  \code{peakShape}, or 
a empty character if you want all.}
}
\value{
plot
}
\description{
ptrSet object
}
\examples{
data(mycobacteriaSet)
plot(mycobacteriaSet)
plot(mycobacteriaSet,typePlot="calibError")
plot(mycobacteriaSet,typePlot="resolution")
plot(mycobacteriaSet,typePlot="peakShape")
}
