% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{plotCalibError}
\alias{plotCalibError}
\title{plot the calibration error bowplot for a ptrSet}
\usage{
plotCalibError(set)
}
\arguments{
\item{set}{ptrSet object}
}
\description{
plot the calibration error bowplot for a ptrSet
}
