% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{plotResolution}
\alias{plotResolution}
\title{plot resolution boxplot for a ptrSet}
\usage{
plotResolution(set)
}
\arguments{
\item{set}{a ptrSet object}
}
\description{
plot resolution boxplot for a ptrSet
}
