% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{ppbConvert}
\alias{ppbConvert}
\title{PPb concentration}
\usage{
ppbConvert(peakList, primaryIon, transmission, U, Td, pd, k = 2 * 10^-9)
}
\arguments{
\item{peakList}{the peak list return by peakList, must contain the isotope 
of primary ion 21.021}

\item{primaryIon}{primary ion quantification in count per extraction}

\item{transmission}{transmission information}

\item{U}{drift voltage}

\item{Td}{drift temperature}

\item{pd}{drift pressure}

\item{k}{reaction rate constant}
}
\value{
a vector with ppb concentration
}
\description{
PPb concentration
}
