% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptairMS-package.R
\docType{package}
\name{ptairMS-package}
\alias{ptairMS}
\alias{ptairMS-package}
\title{ptairMS: Pre-processing PTR-TOF-MS Data}
\description{
This package implements a suite of methods to preprocess data from PTR-TOF-MS instruments (HDF5 format) and generates the 'sample by features' table of peak intensities in addition to the sample and feature metadata (as a singl<e ExpressionSet object for subsequent statistical analysis). This package also permit usefull tools for cohorts management as analyzing data progressively, visualization tools and quality control. The steps include calibration, expiration detection, peak detection and quantification, feature alignment, missing value imputation and feature annotation. Applications to exhaled air and cell culture in headspace are described in the vignettes and examples. This package was used for data analysis of Gassin Delyle study on adults undergoing invasive mechanical ventilation in the intensive care unit due to severe COVID-19 or non-COVID-19 acute respiratory distress syndrome (ARDS), and permit to identfy four potentiel biomarquers of the infection.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/camilleroquencourt/ptairMS/issues}
}

}
\author{
\strong{Maintainer}: camille Roquencourt \email{camille.roquencourt@hotmail.fr}

}
\keyword{internal}
