% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ptrRaw-class}
\alias{ptrRaw-class}
\title{Object contains PTR-TOF-MS raw data frome a file.}
\description{
A ptrRaw object contains PTR-TOF-MS raw data from one h5 file. It is created with the \code{readRaw} function.
}
\section{Slots}{

\describe{
\item{\code{name}}{the file name}

\item{\code{rawM}}{the intensity raw matrix}

\item{\code{mz}}{array of mz axis}

\item{\code{time}}{numeric vector of acquisition time}

\item{\code{calibMzToTof}}{function to convert mz to Tof}

\item{\code{calibToftoMz}}{function to convert tof to mz}

\item{\code{calibCoef}}{calibration coefficients (a,b) such that: mz= ((tof-b)/a)^2}

\item{\code{calibMassRef}}{the reference masses used for the calibration}

\item{\code{calibError}}{the shift error in ppm at the reference masses}

\item{\code{calibSpectr}}{the spectrum of calibration reference masses}

\item{\code{ptrTransmisison}}{matrix with transmission values}

\item{\code{prtReaction}}{a list containing PTR reaction information: drift temperature, pressure and voltage}
}}

