% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinnyApp.R
\name{RunShinnyApp}
\alias{RunShinnyApp}
\title{Graphical interface of ptairMS workflow}
\usage{
RunShinnyApp()
}
\value{
Shinny app
}
\description{
The whole workflow of ptairMS can be run interactively through a graphical user interface,
which provides visualizations (expiration phases, peaks in the raw data, peak table, individual VOCs),
quality controls (calibration, resolution, peak shape and evolution of reagent ions depending on time),
and exploratory data analysis.
}
\examples{
\dontrun{RunShinnyApp()}
}
