% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{sech2}
\alias{sech2}
\title{Fit function Sech2}
\usage{
sech2(p, lf, lr, h, x)
}
\arguments{
\item{p}{peak center}

\item{lf}{lambda 1 , peak width left}

\item{lr}{lambda 2 , peak width right}

\item{h}{peak higth}

\item{x}{vector values}
}
\value{
numeric value
}
\description{
Fit function Sech2
}
