% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_Methods.R
\name{show,ptrRaw-method}
\alias{show,ptrRaw-method}
\title{show a ptrRaw object}
\usage{
\S4method{show}{ptrRaw}(object)
}
\arguments{
\item{object}{a ptrRaw object}
}
\value{
nothing
}
\description{
It indicates the files, the mz range, time acquisition range, and calibration error.
}
