% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{show,ptrSet-method}
\alias{show,ptrSet-method}
\title{show a ptrSet object}
\usage{
\S4method{show}{ptrSet}(object)
}
\arguments{
\item{object}{a ptrSet object}
}
\description{
It indicates the directory, the numer of files taht contain teh directory at the moment, and teh number of processed files.
The two numbers are diffrents, use \code{updatePtrSet} function.
}
