% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{snipBase}
\alias{snipBase}
\title{Baseline estimation}
\usage{
snipBase(sp, widthI = 15, iteI = 3)
}
\arguments{
\item{sp}{an array with spectrum values}

\item{widthI}{width of interval}

\item{iteI}{number of iteration}
}
\value{
baseline estimation of the spectrum
}
\description{
Baseline estimation
}
