% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{timeLimitFunc}
\alias{timeLimitFunc}
\title{Calculate expiration limits on the Chromatogram}
\usage{
timeLimitFunc(TIC, fracMaxTIC = 0.5, minPoints = 2, plotDel = FALSE)
}
\arguments{
\item{TIC}{the TIC vector}

\item{fracMaxTIC}{between 0 and 1. Percentage of the maximum of the Chromatogram amplitude with baseline removal. 
If you want a finer limitation, increase \code{fracMaxTIC}, indeed decrease}

\item{minPoints}{minimum duration of an expiration (in index)}

\item{plotDel}{boolean. If TRUE, the TIC is ploted with limits and threshold.}
}
\value{
a matrix of index, where each colomn correspond to one expriration, the first line its beginning 
and the seconde the end.
}
\description{
Calculate expiration limits on the Chromatogram
}
