% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Raw_Methods.R, R/SetMethods.R
\name{timeLimits}
\alias{timeLimits}
\alias{timeLimits,ptrRaw-method}
\alias{timeLimits,ptrSet-method}
\title{Calculate time limits on the Chromatogram}
\usage{
timeLimits(
  object,
  fracMaxTIC = 0.5,
  traceMasses = NULL,
  minPoints = 2,
  plotDel = FALSE
)

\S4method{timeLimits}{ptrRaw}(
  object,
  fracMaxTIC = 0.5,
  traceMasses = NULL,
  minPoints = 2,
  plotDel = FALSE
)

\S4method{timeLimits}{ptrSet}(
  object,
  fracMaxTIC = 0.5,
  traceMasses = NULL,
  minPoints = 2,
  plotDel = FALSE
)
}
\arguments{
\item{object}{a ptrRaw or ptrSet object}

\item{fracMaxTIC}{between 0 and 1. Percentage of the maximum of the Chromatogram amplitude with baseline removal. 
If you want a finer limitation, increase \code{fracMaxTIC}, indeed decrease}

\item{traceMasses}{NULL or a integer. Correspond to a nominal masses of Extract Ion Chromatogram (EIC)
whose limits you want to compute. If NULL, the limits are calculated on the Total Ion Chromatogram (TIC).}

\item{minPoints}{minimum duration of an expiration (in index).}

\item{plotDel}{boolean. If TRUE, the Chormatogram is ploted with limits and threshold.}
}
\value{
a matrix of index, where each colomn correspond to one expriration, the first row 
it is the beginning and the seconde the end, or NA if no limits are detected.
}
\description{
This function derives limits on the Total Ion Chromatogram TIC, where the intenisty is greater than \code{fracMaxTIC*max(TIC)}, 
where max(TIC)  is the maximum of teh TIC with baseline removal.
In this way,  the expiration limits, or headsapce analysis limits can be detected. So, by setting 
\code{fracMaxTIC} close to 1, the size of teh limits will be restricted.
}
\examples{
library(ptairData)
filePath <- system.file("extdata/exhaledAir/ind1", "ind1-1.h5", package = "ptairData")
raw <- readRaw(filePath)

ind_lim <- timeLimits(raw, fracMaxTIC=0.9, plotDel=TRUE)
ind_lim_acetone <- timeLimits(raw, fracMaxTIC=0.5, traceMasses = 59,plotDel=TRUE)
}
