% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{TransmissionCurve}
\alias{TransmissionCurve}
\title{Estimation of the transmisison curve}
\usage{
TransmissionCurve(x, y)
}
\arguments{
\item{x}{masses}

\item{y}{transmission data}
}
\value{
a numeric vector
}
\description{
Estimation of the transmisison curve
}
