% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{width}
\alias{width}
\title{Calculate the FWHM (Full Width at Half Maximum) in raw data}
\usage{
width(tof, peak, fracMaxTIC = 0.5)
}
\arguments{
\item{tof}{A vector of tof interval}

\item{peak}{A vector of peak Intensity}

\item{fracMaxTIC}{the fraction of the maximum intenisty to compute the width}
}
\value{
the delta FWHM in tof
}
\description{
Calculate the FWHM (Full Width at Half Maximum) in raw data
}
\keyword{internal}
