% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{aggregate}
\alias{aggregate}
\title{aggregate peakgroup for align function}
\usage{
aggregate(subGroupPeak, n.exp)
}
\arguments{
\item{subGroupPeak}{teh group tp aggregate}

\item{n.exp}{number of expiration done in the file}
}
\value{
a matrix with the median of mz, mean of ppb, ppb in background, 
 and percentage of expiration where the peak is detected
 @keywords internal
}
\description{
aggregate peakgroup for align function
}
