% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/alignment.R
\name{alignSamples}
\alias{alignSamples}
\alias{alignSamples,ptrSet-method}
\alias{alignSamples,data.frame-method}
\title{Align mass tables of different samples}
\usage{
alignSamples(
  X,
  ppmGroup = 70,
  fracGroup = 0.8,
  group = NULL,
  pValGreaterThres = 0.005,
  pValLessThres = 0,
  fracExp = 0.3,
  dmzGroup = 0.001,
  ...
)

\S4method{alignSamples}{ptrSet}(
  X,
  ppmGroup = 70,
  fracGroup = 0.8,
  group = NULL,
  pValGreaterThres = 0.001,
  pValLessThres = 0,
  fracExp = 0.3,
  dmzGroup = 0.001,
  quantiUnit = c("ppb", "ncps", "cps")[1],
  bgCorrected = TRUE,
  ...
)

\S4method{alignSamples}{data.frame}(
  X,
  ppmGroup = 70,
  fracGroup = 0.8,
  group = NULL,
  pValGreaterThres = 0.005,
  pValLessThres = 0,
  fracExp = 0.3,
  dmzGroup = 0.005,
  quantiUnit = c("ppb", "ncps", "cps")[1],
  bgCorrected = TRUE,
  sampleMetadata
)
}
\arguments{
\item{X}{ptrSet or data.frame contains peakList with at least column mz, quanti and group}

\item{ppmGroup}{the ppm larger of a group}

\item{fracGroup}{We will keep variables only present in \code{fracGroup} 
percent of at least one \code{group}, if 0 the filter is not applied}

\item{group}{character. A sampleMetadata data colnames. If is \code{NULL},variables not presents
in \code{fracGroup} percent of samples will be deleted. Else, variables not presents
in \code{fracGroup} percent in each group will be deleted.}

\item{pValGreaterThres}{threshold of the he p-value of the unilateral t-test, who test that 
quantidication (in cps) of expiration points are greater than the intenisty of the background.}

\item{pValLessThres}{threshold of the he p-value of the unilateral t-test, who test that 
quantidication (in cps) of expiration points are greater than the intenisty of the background.}

\item{fracExp}{percentage of sample which must have their pvalue less than \code{pValGreaterThres} 
and \code{pValLessThres}}

\item{dmzGroup}{difference of mz of a group for little mz}

\item{...}{not used}

\item{quantiUnit}{ppb, ncps or cps}

\item{bgCorrected}{logical. If the value must be corrected to the background baseline or not.}

\item{sampleMetadata}{if X is a ptrSet, not used, else a data.frame containing sample Metadata}
}
\value{
an expressionSet (Biobase object), with annotaTion in features Data
}
\description{
Align mass tables of different samples
}
\examples{
library(ptairData)
directory <- system.file("extdata/mycobacteria",  package = "ptairData")
dirSet <- createPtrSet(directory,setName="test",mzCalibRef =c(21.022,59.049))
dirSet <- detectPeak(dirSet,mzNominal=c(21,59))
eset <- alignSamples(dirSet,pValGreaterThres=0.05)
Biobase::exprs(eset)
Biobase::fData(eset)
Biobase::pData(eset)
}
