% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/annotation.R
\name{annotateVOC}
\alias{annotateVOC}
\alias{annotateVOC,ExpressionSet-method}
\alias{annotateVOC,data.frame-method}
\alias{annotateVOC,numeric-method}
\title{Putatively annotate VOC mz by using the reference compilation from the literature}
\usage{
annotateVOC(x, ...)

\S4method{annotateVOC}{ExpressionSet}(
  x,
  mzColname = NA,
  ppm = 20,
  prefix = "vocDB_",
  fields = c("mz_Hplus", "formula_Hplus", "cas.name", "reference", "cas", "chebi",
    "chebi.formula", "chebi.monoisotopic.mass", "chebi.name", "chebi.inchi",
    "kegg.compound.id")[1:3],
  ...
)

\S4method{annotateVOC}{data.frame}(x, mzColname = NA, ...)

\S4method{annotateVOC}{numeric}(x, ...)
}
\arguments{
\item{x}{Expression set object (resp. data.frame) (resp. numeric vector) containing
the PTR-MS processed data (resp. containing a column with the mz values) (resp. containing the mz values)}

\item{...}{Additional parameters to be passed to the internal function}

\item{mzColname}{Character: column name from the fData (resp. from the data.frame) containing
the mz values; if set to NA, featureNames from the Expression set (resp. row.names
from the data.frame) will be converted to numerics and used}

\item{ppm}{Numeric: tolerance}

\item{prefix}{Character: prefix for the new 'annotation' columns [default: 'vocDB_']}

\item{fields}{Characer vector: fields of the 'vocDB' database to be queried among:
'mz_Hplus' [default], 'formula_Hplus' [default], 'name' [default], 
reference', 'cas', 'chebi', 'chebi.formula', 'chebi.monoisotopic.mass',
'chebi.name', 'chebi.inchi', 'chebi.kegg.compound.id'}
}
\value{
Returns the Expression set with additional columns in the fData (resp.
the data.frame with additional columns) (resp. a data.frame with columns)
containing the matched 'mz', 'formula', 'name', 'reference', and 'matrix' putative annotations
}
\description{
Putatively annotate VOC mz by using the reference compilation from the literature
}
\examples{
library(ptairMS)
library(ptairData)
directory <- system.file("extdata/mycobacteria",  package = "ptairData")
bacteria.ptrset <- ptairMS::createPtrSet(directory, setName = "bacteria",
mzCalibRef = c(21.022,59.049))
bacteria.ptrset <- ptairMS::detectPeak(bacteria.ptrset)
bacteria.eset <- ptairMS::alignSamples(bacteria.ptrset)
bacteria.eset <- ptairMS::annotateVOC(bacteria.eset)
Biobase::fData(bacteria.eset)
}
