% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{convert_to_mzML}
\alias{convert_to_mzML}
\title{Convert a h5 file to mzML}
\usage{
convert_to_mzML(file)
}
\arguments{
\item{file}{A .h5 file path}
}
\value{
cretae a mzML file in the same directory of the h5 iput file
}
\description{
convert_to_mzML create a mzML file from a h5 file in the same directory with the \code{writeMLData} of
the \code{MSnbase} package
}
