% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_Methods.R
\name{deadTimeCorr}
\alias{deadTimeCorr}
\title{Dead time correction on raw data}
\usage{
deadTimeCorr(raw, ve, vne, r, threshold = 0.1)
}
\arguments{
\item{raw}{ptrRaw object}

\item{ve}{exenting dead time}

\item{vne}{non extending dead time}

\item{r}{number of extraction}

\item{threshold}{only bin of intenisty more then threshold*r which be corrected}
}
\value{
a ptrRaw object with the raw matrix corrected
}
\description{
Dead time correction on raw data
}
