% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{determinePeakShape}
\alias{determinePeakShape}
\title{Determine peak shape from raw data in tof}
\usage{
determinePeakShape(raw, plotShape = FALSE)
}
\arguments{
\item{raw}{a \code{\link[ptairMS]{ptrRaw-class}} object}

\item{plotShape}{if true plot each reference peak and the average peak (the peak shape)}
}
\value{
A list of two vectors which are the reference peak normalized tof and intensity.
}
\description{
This function use the method descibe by average and al 2013, for determine a peak 
shape from the raw data : \cr 
$peak_i(Delta_i,A_i,t_i) = interpolation(x= tof.ref * Delta_i + t_i,y = A_i * 
peak.ref, xout= TOF_i )$ where peak.ref and tof.ref are peaks reference use for mass calibration.
}
