% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{exportSampleMetada}
\alias{exportSampleMetada}
\title{export sampleMetadata}
\usage{
exportSampleMetada(set, saveFile)
}
\arguments{
\item{set}{a ptrSet object}

\item{saveFile}{a file path in tsv extension where the data.frame will be 
exported}
}
\value{
nothing
}
\description{
export sampleMetadata
}
\examples{
dirRaw <- system.file("extdata/exhaledAir", package = "ptairData")
exhaledPtrset <- createPtrSet(dir=dirRaw, setName="exhaledPtrset", 
mzCalibRef = c(21.022, 60.0525), fracMaxTIC = 0.7, saveDir = NULL )
saveFile<-file.path(getwd(),'sampleMetadata.tsv')
#exportSampleMetada(exhaledPtrset ,saveFile)
}
