% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{extractEIC}
\alias{extractEIC}
\title{extract all raw EIC from a pre-definied peak List}
\usage{
extractEIC(raw, peak, peakQuantil = 0.01, fctFit = "sech2")
}
\arguments{
\item{raw}{ptrRaw object}

\item{peak}{a data.frame with a column named 'Mz'. The Mz of the VOC detected}

\item{peakQuantil}{the quantile of the peak shape to determine the borne of 
the EIC}

\item{fctFit}{function used to fit peak}
}
\value{
list containing all EIC and the mz borne for all peak
}
\description{
extract all raw EIC from a pre-definied peak List
}
\keyword{internal}
