% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{getDirectory}
\alias{getDirectory}
\title{get the files diretory of a ptrSet}
\usage{
getDirectory(ptrSet)
}
\arguments{
\item{ptrSet}{ptrSte object}
}
\value{
the directory in absolute path as character
}
\description{
get the files diretory of a ptrSet
}
\examples{
data(exhaledPtrset )
getDirectory(exhaledPtrset )
}
