% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{getPeakList}
\alias{getPeakList}
\title{get the peak list of a ptrSet object}
\usage{
getPeakList(set)
}
\arguments{
\item{set}{ptrSet object}
}
\value{
a list containing: 
\itemize{ 
\item raw: for each files a list of peak list for each time periods and background, if \code{fracMaxTic} is zero
\item aligned: for each file the peak List after aligning between time periods and removing background threshold}
}
\description{
get the peak list of a ptrSet object
}
\examples{
library(ptairData)
directory <- system.file("extdata/mycobacteria",  package = "ptairData")
dirSet <- createPtrSet(directory,setName="test",mzCalibRef=c(21.022,59.049))
dirSet <- detectPeak(dirSet , mzNominal=c(59,60))
getPeakList(dirSet)$aligned
getPeakList(dirSet)$raw
}
