% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{getSampleMetadata}
\alias{getSampleMetadata}
\title{get sampleMetadata of a ptrSet}
\usage{
getSampleMetadata(set)
}
\arguments{
\item{set}{a ptrSet object}
}
\value{
a data.frame
}
\description{
get sampleMetadata of a ptrSet
}
\examples{
dirRaw <- system.file("extdata/exhaledAir", package = "ptairData")
exhaledPtrset <- createPtrSet(dir=dirRaw, setName="exhaledPtrset", 
mzCalibRef = c(21.022, 60.0525), fracMaxTIC = 0.7, saveDir = NULL)
SMD<-getSampleMetadata(exhaledPtrset)
}
