% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{impute}
\alias{impute}
\title{Impute missing values of an expression set from raw data}
\usage{
impute(eSet, ptrSet, parallelize = FALSE, nbCores = 2)
}
\arguments{
\item{eSet}{an expression set return by \code{\link[ptairMS]{alignSamples}} 
function}

\item{ptrSet}{a \code{\link[ptairMS]{ptrSet-class}} object processed by 
\code{\link[ptairMS]{detectPeak}} function}

\item{parallelize}{boolean. If \code{TRUE} loop over files will be paralleled}

\item{nbCores}{number of cluster to use for parallel computation.}
}
\value{
the same expression set as in input, with missing values imputing
}
\description{
Imputing missing values by returning back to the raw data and fitting the 
peak shape function on the noise or residuals signals if peaks already detected.
}
\examples{
data(exhaledPtrset)
getSampleMetadata(exhaledPtrset)
eSet <- alignSamples(exhaledPtrset,pValGreaterThres=0.05,fracGroup=0)
Biobase::exprs(eSet)
eSet <- impute(eSet,exhaledPtrset)
Biobase::exprs(eSet)
}
