% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{imputeMat}
\alias{imputeMat}
\title{Impute missing values on an matrix set from an ptrSet}
\usage{
imputeMat(X, ptrSet, quantiUnit)
}
\arguments{
\item{ptrSet}{processed by detectPeak function}

\item{eSet}{an expression set retun by alignSamples function}
}
\value{
the same expression set as in input, with missing values imputing
}
\description{
Imputing missing values by returning back to the raw data and fitting the 
peak shape function on the noise / residuals
}
