% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mycobacteriaSet}
\alias{mycobacteriaSet}
\title{PtrSet of the mycobacteria rdhf 5 files}
\format{
a ptrSet object
}
\usage{
mycobacteriaSet
}
\description{
a ptrSet object, created by the \code{ptairMS::createPtrSet} function, 
containing processed mycobacteria dataset (files present in the 
inst/extdata/mycobacteria) by applying \code{ptairMS::detectPeak} function
}
\keyword{datasets}
