% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{plot,ptrSet-method}
\alias{plot,ptrSet-method}
\alias{plot.ptrSet}
\title{ptrSet object}
\usage{
\S4method{plot}{ptrSet}(x, y, typePlot = "")
}
\arguments{
\item{x}{a ptrSet object}

\item{y}{not use}

\item{typePlot}{could be : \code{calibError}, \code{resolution},  \code{peakShape}, or 
a empty character if you want all.}
}
\value{
plot
}
\description{
ptrSet object
}
\examples{
library(ptairData)
directory <- system.file("extdata/mycobacteria",  package = "ptairData")
mycobacteria <- createPtrSet(dir= directory, setName="mycobacteria",mzCalibRef= c(21.022, 59.049141))
plot(mycobacteria)
plot(mycobacteria,typePlot="calibError")
plot(mycobacteria,typePlot="resolution")
plot(mycobacteria,typePlot="peakShape")
}
