% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{plotPeakShape}
\alias{plotPeakShape}
\title{plot the average peak shape of reference calibration masses for a ptrSet}
\usage{
plotPeakShape(set, showAverage = FALSE)
}
\arguments{
\item{set}{ptrSet object}

\item{showAverage}{boolean}
}
\value{
ggplot object
}
\description{
plot the average peak shape of reference calibration masses for a ptrSet
}
