% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ptrSet-class}
\alias{ptrSet-class}
\title{Object contains a set of PTR-TOF-MS raw data}
\description{
A ptrSet object is related to a directory that contains several PTR-TOF-MS raw data in h5 format. It is created 
with the \code{createPtrSet} function.
}
\section{Slots}{

\describe{
\item{\code{parameter}}{the input parameters value of the function createPtrSet}

\item{\code{sampleMetadata}}{data frame of sample metadata, with file names in row names}

\item{\code{mzCalibRef}}{the masses uses for calibration for each file}

\item{\code{signalCalibRef}}{the spectrum of mass calibration for each file}

\item{\code{errorCalibPpm}}{the calibration error for each file}

\item{\code{coefCalib}}{the coefficients of mass axis calibration for each file}

\item{\code{primaryIon}}{the quantity in count per acquisition time of the isotope of primary ion (H30_18, mz 21.022) for each file}

\item{\code{resolution}}{estimation of the resolution for each file based on the calibration reference masses}

\item{\code{TIC}}{the TIC for each file}

\item{\code{breathTracer}}{EIC for expirations/head spaces detection}

\item{\code{timeLimit}}{the index of time limit for each file}

\item{\code{peakListRaw}}{the peak list before a potential aggregation  of expirations}

\item{\code{peakListAligned}}{the peakList with expirations aggregation  for each file}
}}

