% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{resetSampleMetadata}
\alias{resetSampleMetadata}
\title{reset the default sampleMetadata, containing the suborders names and the 
acquisition dates as columns.}
\usage{
resetSampleMetadata(ptrset)
}
\arguments{
\item{ptrset}{a ptrser object}
}
\value{
a data.frame
}
\description{
reset the default sampleMetadata, containing the suborders names and the 
acquisition dates as columns.
}
\examples{
data(exhaledPtrset )
SMD<- resetSampleMetadata(exhaledPtrset )
}
