% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{resetSampleMetadata}
\alias{resetSampleMetadata}
\title{reset the default sampleMetadata}
\usage{
resetSampleMetadata(ptrset)
}
\arguments{
\item{ptrset}{a ptrser object}
}
\value{
a data.frame
}
\description{
reset the default sampleMetadata
}
\examples{
library(ptairData)
directory <- system.file("extdata/mycobacteria",  package = "ptairData")
mycobacteria <- createPtrSet(dir= directory, setName="mycobacteria",mzCalibRef= c(21.022, 59.049141))
SMD<- resetSampleMetadata(mycobacteria)
}
