% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{rmPeakList}
\alias{rmPeakList}
\title{remove the peakList of an ptrSet object}
\usage{
rmPeakList(object)
}
\arguments{
\item{object}{ptrSet object}
}
\value{
a ptrSet
}
\description{
This function is usefull when you want to change the parameters of the detect 
peak function. First delete the peakLIst with \code{rmPeakList}, and apply \code{detectPeak}
with the new parameters.
}
\examples{
library(ptairMS)
data(mycobacteriaSet)
mycobacteriaSet<-rmPeakList(mycobacteriaSet)
}
