% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{setSampleMetadata}
\alias{setSampleMetadata}
\title{set sampleMetadata in a ptrSet}
\usage{
setSampleMetadata(set, sampleMetadata)
}
\arguments{
\item{set}{a ptrSet object}

\item{sampleMetadata}{a data.frame with all file names of the ptrSet in row names}
}
\value{
the ptrSet object in argument with the sampleMetadata modified
}
\description{
set sampleMetadata in a ptrSet
}
\examples{
library(ptairMS)
data(mycobacteriaSet)
SMD<-getSampleMetadata(mycobacteriaSet)
colnames(SMD)[1]<-"species"
mycobacteria<-setSampleMetadata(mycobacteriaSet,SMD)
}
