% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Raw_Methods.R, R/SetMethods.R
\name{timeLimits}
\alias{timeLimits}
\alias{timeLimits,ptrRaw-method}
\alias{timeLimits,ptrSet-method}
\title{Calculate time limits on the Chromatogram}
\usage{
timeLimits(
  object,
  fracMaxTIC = 0.5,
  fracMaxTICBg = 0.5,
  derivThresholdExp = 0.5,
  derivThresholdBg = 0.01,
  traceMasses = NULL,
  minPoints = 2,
  degreeBaseline = 1,
  plotDel = FALSE
)

\S4method{timeLimits}{ptrRaw}(
  object,
  fracMaxTIC = 0.5,
  fracMaxTICBg = 0.5,
  derivThresholdExp = 0.5,
  derivThresholdBg = 0.01,
  traceMasses = NULL,
  minPoints = 2,
  degreeBaseline = 1,
  plotDel = FALSE
)

\S4method{timeLimits}{ptrSet}(
  object,
  fracMaxTIC = 0.5,
  fracMaxTICBg = 0.5,
  derivThresholdExp = 0.5,
  derivThresholdBg = 0.01,
  minPoints = 2,
  degreeBaseline = 1,
  plotDel = FALSE
)
}
\arguments{
\item{object}{a ptrRaw or ptrSet object}

\item{fracMaxTIC}{between 0 and 1. Percentage of the maximum of the Chromatogram amplitude with baseline 
removal. If you want a finer limitation, increase \code{fracMaxTIC}, indeed decrease}

\item{fracMaxTICBg}{same as fracMaxTIC but for backgroud dettcion (lower than fracMaxTIC*max(TIC))}

\item{derivThresholdExp}{the tresphold of the difference between two succesive points of the expiration}

\item{derivThresholdBg}{the tresphold of the difference between two succesive points of the background}

\item{traceMasses}{NULL or a integer. Correspond to a nominal masses of Extract Ion Chromatogram (EIC)
whose limits you want to compute. If NULL, the limits are calculated on the Total Ion Chromatogram (TIC).}

\item{minPoints}{minimum duration of an expiration (in index).}

\item{degreeBaseline}{the degree of polynomial baseline function}

\item{plotDel}{boolean. If TRUE, the Chormatogram is ploted with limits and threshold.}
}
\value{
a list with expirations limits (a matrix of index, where each colomn correspond to one expriration, the first row 
it is the beginning and the seconde the end, or NA if no limits are detected) and index of the background.
}
\description{
This function derives limits on the Total Ion Chromatogram TIC, where the intenisty is greater than \code{fracMaxTIC*max(TIC)}, 
where max(TIC)  is the maximum of teh TIC with baseline removal.
In this way,  the expiration limits, or headsapce analysis limits can be detected. So, by setting 
\code{fracMaxTIC} close to 1, the size of teh limits will be restricted. This function also detemine 
the index corresponding to the backgound, where variation between two succesive point can be control with 
derivThreshold paramter.
}
\examples{
library(ptairData)
filePath <- system.file("extdata/exhaledAir/ind1", "ind1-1.h5", package = "ptairData")
raw <- readRaw(filePath)

timLim <- timeLimits(raw, fracMaxTIC=0.9, plotDel=TRUE)
timLim_acetone <- timeLimits(raw, fracMaxTIC=0.5, traceMasses = 59,plotDel=TRUE)
}
