% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/writeEset.R
\name{writeEset}
\alias{writeEset}
\alias{writeEset,ExpressionSet-method}
\title{Exporting an ExpressionSet instance into 3 tabulated files 'dataMatrix.tsv',
sampleMetadata.tsv', and 'variableMetadata.tsv'}
\usage{
writeEset(x, dirName, overwrite = FALSE, verbose = TRUE)

\S4method{writeEset}{ExpressionSet}(x, dirName, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{An S4 object of class \code{ExpressionSet}}

\item{dirName}{Character: directory where the tables should be written}

\item{overwrite}{Logical: should existing files be overwritten?}

\item{verbose}{Logical: should messages be printed?}
}
\value{
No object returned.
}
\description{
Note that the \code{dataMatrix} is transposed before
export (e.g., the samples are written column wise in the 'dataMatrix.tsv'
exported file).
}
\examples{
data(exhaledPtrset )
eset <- ptairMS::alignSamples(exhaledPtrset ) 
\dontrun{
writeEset(eset, dirName = file.path(getwd(), "processed_dataset"))
}
}
